+++ client/client.go
@@ -33,7 +33,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	driver "github.com/arangodb/go-driver"
+	driver "github.com/arangodb/go-driver/v2/arangodb"
 
 	"github.com/arangodb-helper/arangodb/pkg/api"
 )