+++ service/peer.go
@@ -26,7 +26,7 @@ import (
 	"strconv"
 	"strings"
 
-	driver "github.com/arangodb/go-driver"
+	driver "github.com/arangodb/go-driver/v2/arangodb"
 
 	"github.com/arangodb-helper/arangodb/pkg/definitions"
 )