+++ service/server.go
@@ -35,7 +35,7 @@ import (
 
 	"github.com/rs/zerolog"
 
-	driver "github.com/arangodb/go-driver"
+	driver "github.com/arangodb/go-driver/v2/arangodb"
 
 	"github.com/arangodb-helper/arangodb/client"
 	"github.com/arangodb-helper/arangodb/pkg/definitions"